refactor: change errorboundary structure
