+++ src/App.tsx
@@ -1,4 +1,4 @@
-import React, { Component, ErrorInfo, ReactNode } from 'react';
+import { Component, ChangeEvent } from 'react';
 import { Repository } from './types/ServerAnswer.type';
 
 type AppState = {
@@ -7,47 +7,6 @@ type AppState = {
   loading: boolean;
   error: boolean;
 };
-
-type ErrorBoundaryState = {
-  hasError: boolean;
-};
-
-type ErrorBoundaryProps = {
-  children: ReactNode;
-};
-
-class ErrorBoundary extends Component<ErrorBoundaryProps, ErrorBoundaryState> {
-  constructor(props: ErrorBoundaryProps) {
-    super(props);
-    this.state = { hasError: false };
-  }
-
-  static getDerivedStateFromError() {
-    return { hasError: true };
-  }
-
-  componentDidCatch(error: Error, errorInfo: ErrorInfo) {
-    console.error('Error caught in ErrorBoundary:', error, errorInfo);
-  }
-
-  render() {
-    if (this.state.hasError) {
-      return (
-        <h1>
-          Something went wrong.
-          <button
-            className="btn-reset primary-btn"
-            onClick={() => this.setState({ hasError: false })}
-          >
-            Try again
-          </button>
-        </h1>
-      );
-    }
-
-    return this.props.children;
-  }
-}
 
 class App extends Component<object, AppState> {
   constructor(props: object) {
@@ -100,7 +59,7 @@ class App extends Component<object, AppState> {
     this.fetchResults(searchRepo);
   };
 
-  handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+  handleInputChange = (event: ChangeEvent<HTMLInputElement>) => {
     this.setState({ searchRepo: event.target.value });
   };
 
@@ -108,56 +67,54 @@ class App extends Component<object, AppState> {
     const { searchRepo, results, loading, error } = this.state;
 
     return (
-      <ErrorBoundary>
-        <main className="app">
-          <div className="app__info">
-            <h1 className="app__title">Search for repositories in GitHub 👨‍💻</h1>
-            <p className="app__descr">
-              Enter a name and get the first 50 results in the format: "Account
-              Name/Repository".
-            </p>
-          </div>
-          <search className="app__search">
-            <form className="app__form">
-              <input
-                className="form-field__input input-reset"
-                placeholder="Enter a repository"
-                type="text"
-                value={searchRepo}
-                onChange={this.handleInputChange}
-              />
-              <button
-                onClick={this.handleSearch}
-                className="primary-btn btn-reset"
-              >
-                Search
-              </button>
-            </form>
-          </search>
-          <div className="app__results">
-            {loading && (
-              <div className="loader-container">
-                <div className="loader"></div>
-              </div>
-            )}
-            {error ? (
-              <p className="app__error">Error loading results. 😿</p>
-            ) : (
-              <ul className="app__list list-reset">
-                {results && results.length > 0 ? (
-                  results.map((result) => (
-                    <li className="app__item" key={result.id}>
-                      {result.full_name}
-                    </li>
-                  ))
-                ) : (
-                  <li>No results found 🥺</li>
-                )}
-              </ul>
-            )}
-          </div>
-        </main>
-      </ErrorBoundary>
+      <main className="app">
+        <div className="app__info">
+          <h1 className="app__title">Search for repositories in GitHub 👨‍💻</h1>
+          <p className="app__descr">
+            Enter a name and get the first 50 results in the format: "Account
+            Name/Repository".
+          </p>
+        </div>
+        <search className="app__search">
+          <form className="app__form">
+            <input
+              className="form-field__input input-reset"
+              placeholder="Enter a repository"
+              type="text"
+              value={searchRepo}
+              onChange={this.handleInputChange}
+            />
+            <button
+              onClick={this.handleSearch}
+              className="primary-btn btn-reset"
+            >
+              Search
+            </button>
+          </form>
+        </search>
+        <div className="app__results">
+          {loading && (
+            <div className="loader-container">
+              <div className="loader"></div>
+            </div>
+          )}
+          {error ? (
+            <p className="app__error">Error loading results. 😿</p>
+          ) : (
+            <ul className="app__list list-reset">
+              {results && results.length > 0 ? (
+                results.map((result) => (
+                  <li className="app__item" key={result.id}>
+                    {result.full_name}
+                  </li>
+                ))
+              ) : (
+                <li>No results found 🥺</li>
+              )}
+            </ul>
+          )}
+        </div>
+      </main>
     );
   }
 }
